added ability for squads to merge after retreating!
moved code around and refactored substantially.

All higher-level squad functions are now in SquadControl.lua.
Squad-related things (trimSquad, addMemberToSquad, etc.) are in Squad.lua.
Probably still need to move some things out of robotarmyhelpers.lua into Squad.lua.
