+++ robolib/util.lua
@@ -2,6 +2,17 @@ require("stdlib/game")
 require("stdlib/log/logger")
 
 
+--examines the given table, and if it finds a nil element it will remove it
+--from the table.
+function removeNilsFromTable(tableIN)
+    for i, element in pairs(tableIN) do
+        if element == nil then
+            table.remove(tableIN, i)
+        end
+    end
+end
+
+
 function table.contains(table, element)
     for _, value in pairs(table) do
         if value == element then
@@ -28,7 +39,6 @@ function table.countValidElements(inputTable)
     local count = 0
     for key, element in pairs(inputTable) do
         if element and key ~= "size" then --avoids issue of squad member table contains an additional entry for size
-
             if element.valid then
                 count = count + 1
             end
